add authorization & feature test for visitor access file pages

+++ app/Http/Controllers/FileController.php
@@ -3,10 +3,10 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\StoreFileRequest;
-use App\Http\Requests\UpdateFileRequest;
 use App\Models\File;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Storage;
 use Sqids\Sqids;
 use Illuminate\Support\Str;
@@ -51,6 +51,7 @@ class FileController extends Controller
      */
     public function show(File $file)
     {
+        Gate::authorize('mustOwner', $file);
         return view('pages.home.file', compact('file'));
     }
 
@@ -59,6 +60,7 @@ class FileController extends Controller
      */
     public function destroy(File $file)
     {
+        Gate::authorize('mustOwner', $file);
         if (Storage::exists($file->storage_path)) Storage::delete($file->storage_path);
         $file->delete();
         return back();
